refactor: move test utility SearchParamsUpdater into shared utility
